Version 1

+++ src/main/java/com/example/YouOnWork/API/Model/User.java
@@ -6,12 +6,12 @@ import javax.persistence.*;
 @Table(name = "user")
 public class User {
 
-    private Long id ;
+    private int id ;
     private String username;
 //    @Column( columnDefinition = "LONGVARBINARY")
     private String password;
 
-    public User(Long id, String username, String password) {
+    public User(int id, String username, String password) {
         this.id = id;
         this.username = username;
         this.password = password;
@@ -28,11 +28,11 @@ public class User {
     @Id
     @Column(name = "id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    public Long getId() {
+    public int getId() {
         return id;
     }
 
-    public void setId(Long id) {
+    public void setId(int id) {
         this.id = id;
     }
